[python] Support table reader for PyPaimon

+++ paimon-python/pypaimon/tests/reader_append_only_test.py
@@ -37,7 +37,7 @@ class AoReaderTest(unittest.TestCase):
 
         cls.pa_schema = pa.schema([
             ('user_id', pa.int32()),
-            ('item_id', pa.int32()),
+            ('item_id', pa.int64()),
             ('behavior', pa.string()),
             ('dt', pa.string())
         ])
